requirements.txt
streamlit
langchain
openai
faiss-cpu
PyPDF2
sentence-transformers
huggingface-hub
python-dotenv
tiktoken
